[ui] SplashActivity UI 구현
상태바, 네비게이션바는 블랙(#000000) 으로 변경하였음.
