+++ app/src/main/java/com/jangjh123/allpouse_android/ui/splash/SplashActivity.kt
@@ -1,0 +1,53 @@
+package com.jangjh123.allpouse_android.ui.splash
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.tooling.preview.Preview
+import com.jangjh123.allpouse_android.ui.theme.AllPouseAndroidTheme
+import com.jangjh123.allpouse_android.ui.theme.mainColor
+import com.jangjh123.allpouse_android.ui.theme.mainGradient0
+import com.jangjh123.allpouse_android.ui.theme.mainGradient1
+
+class SplashActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            AllPouseAndroidTheme {
+                SplashActivityContent()
+            }
+        }
+    }
+}
+
+@Composable
+private fun SplashActivityContent() {
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(
+                brush = Brush.verticalGradient(
+                    colors = listOf(
+                        mainColor(),
+                        mainGradient0(),
+                        mainGradient1()
+                    )
+                )
+            )
+    )
+}
+
+
+@Preview(showBackground = true)
+@Composable
+private fun DefaultPreview() {
+    AllPouseAndroidTheme {
+        SplashActivityContent()
+    }
+}
