+++ app/src/main/res/values/themes.xml
@@ -2,6 +2,7 @@
 <resources>
 
     <style name="Theme.AllPouseAndroid" parent="android:Theme.Material.Light.NoActionBar">
-        <item name="android:statusBarColor">@color/purple_700</item>
+        <item name="android:statusBarColor">#000000</item>
+        <item name="android:navigationBarColor">#000000</item>
     </style>
 </resources>
